cyber-sensei/backend/requirements.txt
# backend/requirements.txt

# Web Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
alembic
psycopg2-binary

# Caching & Task Queue
redis
celery

# AI & Vector Store
langchain
langchain-core
langchain-openai
langchain-community
chromadb
sentence-transformers
openai-whisper
ffmpeg-python

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
PyJWT
email-validator

# Machine Learning
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Logging & Monitoring
elasticsearch>=8.0.0
python-json-logger>=2.0.0

# Utils
python-dotenv
pydantic>=2.0.0
python-multipart
aiofiles
pypdf
python-docx
slowapi
python-magic
python-magic-bin; sys_platform == 'win32'

# Testing
pytest
pytest-asyncio
httpx
email-validator
python-dateutil>=2.8.0
